refactor exception
